server/requirements.txt
elasticsearch
kafka-python